add comment, minor refactoring in tests
